requirements.txt
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
requests>=2.31.0,<3.0.0
plotly>=5.20.0,<6.0.0
dash>=2.16.0,<3.0.0
dash-bootstrap-components>=1.6.0,<2.0.0
folium>=0.16.0,<1.0.0
geopandas>=0.14.0,<1.0.0
shapely>=2.0.0,<3.0.0
scikit-learn>=1.4.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
pytest>=8.0.0,<9.0.0
psycopg2-binary>=2.9.0,<3.0.0
